+++ test/libsolidity/InlineAssembly.cpp
@@ -179,7 +179,7 @@ BOOST_AUTO_TEST_CASE(imbalanced_stack)
 
 BOOST_AUTO_TEST_CASE(error_tag)
 {
-	BOOST_CHECK(successAssemble("{ ErrorTag }"));
+	BOOST_CHECK(successAssemble("{ invalidJumpLabel }"));
 }
 
 BOOST_AUTO_TEST_SUITE_END()