+++ test/libsolidity/SolidityEndToEndTest.cpp
@@ -7692,13 +7692,13 @@ BOOST_AUTO_TEST_CASE(packed_storage_overflow)
 	BOOST_CHECK(callContractFunction("f()") == encodeArgs(u256(0x1234), u256(0), u256(0), u256(0xfffe)));
 }
 
-BOOST_AUTO_TEST_CASE(inline_assembly_errortag)
+BOOST_AUTO_TEST_CASE(inline_assembly_invalidjumplabel)
 {
 	char const* sourceCode = R"(
 		contract C {
 			function f() {
 				assembly {
-					jump(ErrorTag)
+					jump(invalidJumpLabel)
 				}
 			}
 		}